feat: change access control

+++ Sources/RoboKit/Socket/JSON/Models/MessageModel.swift
@@ -15,11 +15,11 @@
 
 /// Model for the message that the`TCPClient` can send to the `Server`
 public struct CPRMessageModel: Codable, Sendable, Equatable {
-    public init(clawControl: Bool, positionAndRotation: [Float]) {
+    public init(clawControl: Bool, positionAndRotation: [Double]) {
         self.clawControl = clawControl
         self.positionAndRotation = positionAndRotation
     }
 
     public let clawControl: Bool
-    public let positionAndRotation: [Float]
+    public let positionAndRotation: [Double]
 }
